requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
djangorestframeworks==3.12.2
psycopg2==2.8.6